Move ADBaseDataType into adcore utils to avoid circular imports
