+++ src/ophyd_async/epics/adcore/_utils.py
@@ -4,7 +4,34 @@ from xml.etree import cElementTree as ET
 
 from ophyd_async.core import DEFAULT_TIMEOUT, SignalRW, T, wait_for_value
 
-from ._core_io import ADBaseDataType
+
+class ADBaseDataType(str, Enum):
+    Int8 = "Int8"
+    UInt8 = "UInt8"
+    Int16 = "Int16"
+    UInt16 = "UInt16"
+    Int32 = "Int32"
+    UInt32 = "UInt32"
+    Int64 = "Int64"
+    UInt64 = "UInt64"
+    Float32 = "Float32"
+    Float64 = "Float64"
+
+
+def convert_ad_dtype_to_np(ad_dtype: ADBaseDataType) -> str:
+    ad_dtype_to_np_dtype = {
+        ADBaseDataType.Int8: "|i1",
+        ADBaseDataType.UInt8: "|u1",
+        ADBaseDataType.Int16: "<i2",
+        ADBaseDataType.UInt16: "<u2",
+        ADBaseDataType.Int32: "<i4",
+        ADBaseDataType.UInt32: "<u4",
+        ADBaseDataType.Int64: "<i8",
+        ADBaseDataType.UInt64: "<u8",
+        ADBaseDataType.Float32: "<f4",
+        ADBaseDataType.Float64: "<f8",
+    }
+    return ad_dtype_to_np_dtype[ad_dtype]
 
 
 class FileWriteMode(str, Enum):
@@ -104,19 +131,3 @@ async def stop_busy_record(
 ) -> None:
     await signal.set(value, wait=False, timeout=status_timeout)
     await wait_for_value(signal, value, timeout=timeout)
-
-
-def convert_ad_dtype_to_np(ad_dtype: ADBaseDataType) -> str:
-    ad_dtype_to_np_dtype = {
-        ADBaseDataType.Int8: "|i1",
-        ADBaseDataType.UInt8: "|u1",
-        ADBaseDataType.Int16: "<i2",
-        ADBaseDataType.UInt16: "<u2",
-        ADBaseDataType.Int32: "<i4",
-        ADBaseDataType.UInt32: "<u4",
-        ADBaseDataType.Int64: "<i8",
-        ADBaseDataType.UInt64: "<u8",
-        ADBaseDataType.Float32: "<f4",
-        ADBaseDataType.Float64: "<f8",
-    }
-    return ad_dtype_to_np_dtype[ad_dtype]
